Base class changes
