+++ custom_components/sonoff/sonoff_local.py
@@ -5,12 +5,12 @@ from base64 import b64encode, b64decode
 from typing import Callable, List
 
 from aiohttp import ClientSession
-from zeroconf import ServiceBrowser, Zeroconf, ServiceStateChange
 
 from Crypto.Cipher import AES
 from Crypto.Hash import MD5
 from Crypto.Random import get_random_bytes
 from Crypto.Util.Padding import pad, unpad
+from zeroconf import ServiceBrowser, Zeroconf, ServiceStateChange
 
 _LOGGER = logging.getLogger(__name__)
